Distance RMSD Potential interaction function
Implemented the new interaction function type as well
as the handling in the topology file. Distance RMSD
Potentials can now be defined in the topology file in 
this way:

[ drmsd_restraints ]
;   ai      aj      type    d0
    9       39      1       0.5
    19      49      1       0.5
    29      59      1       0.5

Atom pair ai, aj which forms a distance pair. Default 
function type 1. d0 is the reference distance of this 
atom pair.