Distance RMSD potential replica exchange
Deliver a lambda scaling functionality to the distance based RMSD
potential. Both the reference rmsd as well as the reference distances
can be scaled from state A to B. The references drmsd values are given
by the .mdp parameters drmsd-ref and drmsd-refB. The reference distances
are specified in the topology file

[ drmsd_restraints ]
;   ai  aj  type  d0     d0B
    9   39  1     0.5    0.6
    19  49  1     0.5    0.6
    29  59  1     0.5    0.6

For detailed information see the documentation.